Move build_table to cli
