+++ crates/cli/src/subcommands/sql.rs
@@ -9,9 +9,8 @@ use crate::util::{database_identity, get_auth_header, ResponseExt, UNSTABLE_WARN
 use anyhow::Context;
 use clap::{Arg, ArgAction, ArgMatches};
 use reqwest::RequestBuilder;
-use spacetimedb::sql::compiler::build_table;
 use spacetimedb_lib::de::serde::SeedWrapper;
-use spacetimedb_lib::sats::Typespace;
+use spacetimedb_lib::sats::{satn, ProductType, ProductValue, Typespace};
 
 pub fn cli() -> clap::Command {
     clap::Command::new("sql")
@@ -188,14 +187,51 @@ pub async fn exec(config: Config, args: &ArgMatches) -> Result<(), anyhow::Error
     Ok(())
 }
 
+/// Generates a [`tabled::Table`] from a schema and rows, using the style of a psql table.
+fn build_table<E>(
+    schema: &ProductType,
+    rows: impl Iterator<Item = Result<ProductValue, E>>,
+) -> Result<tabled::Table, E> {
+    let mut builder = tabled::builder::Builder::default();
+    builder.set_header(
+        schema
+            .elements
+            .iter()
+            .enumerate()
+            .map(|(i, e)| e.name.clone().unwrap_or_else(|| format!("column {i}").into())),
+    );
+
+    let ty = Typespace::EMPTY.with_type(schema);
+    for row in rows {
+        let row = row?;
+        builder.push_record(ty.with_values(&row).enumerate().map(|(idx, value)| {
+            let ty = satn::PsqlType {
+                tuple: ty.ty(),
+                field: &ty.ty().elements[idx],
+                idx,
+            };
+
+            satn::PsqlWrapper { ty, value }.to_string()
+        }));
+    }
+
+    let mut table = builder.build();
+    table.with(tabled::settings::Style::psql());
+
+    Ok(table)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::api::StmtStatsJson;
     use itertools::Itertools;
     use serde_json::value::RawValue;
-    use spacetimedb_lib::sats::ProductType;
-    use spacetimedb_lib::{AlgebraicType, AlgebraicValue};
+    use spacetimedb_lib::error::ResultTest;
+    use spacetimedb_lib::sats::time_duration::TimeDuration;
+    use spacetimedb_lib::sats::timestamp::Timestamp;
+    use spacetimedb_lib::sats::{product, GroundSpacetimeType, ProductType};
+    use spacetimedb_lib::{AlgebraicType, AlgebraicValue, ConnectionId, Identity};
 
     fn make_row(row: &[AlgebraicValue]) -> Result<Box<RawValue>, serde_json::Error> {
         let json = serde_json::json!(row);
@@ -401,6 +437,105 @@ Roundtrip time: 1.00ms"#,
 (1 row) [inserted: 1, deleted: 1, updated: 1, server: 1.00ms]
 Roundtrip time: 1.00ms"#,
         )?;
+
+        Ok(())
+    }
+
+    fn expect_psql_table(ty: &ProductType, rows: Vec<ProductValue>, expected: &str) {
+        let table = build_table(ty, rows.into_iter().map(Ok::<_, ()>)).unwrap().to_string();
+        let mut table = table.split('\n').map(|x| x.trim_end()).join("\n");
+        table.insert(0, '\n');
+        assert_eq!(expected, table);
+    }
+
+    // Verify the output of `sql` matches the inputs that return true for [`AlgebraicType::is_special()`]
+    #[test]
+    fn output_special_types() -> ResultTest<()> {
+        // Check tuples
+        let kind: ProductType = [
+            AlgebraicType::String,
+            AlgebraicType::U256,
+            Identity::get_type(),
+            ConnectionId::get_type(),
+            Timestamp::get_type(),
+            TimeDuration::get_type(),
+        ]
+        .into();
+        let value = product![
+            "a",
+            Identity::ZERO.to_u256(),
+            Identity::ZERO,
+            ConnectionId::ZERO,
+            Timestamp::UNIX_EPOCH,
+            TimeDuration::ZERO
+        ];
+
+        expect_psql_table(
+            &kind,
+            vec![value],
+            r#"
+ column 0 | column 1 | column 2                                                           | column 3                           | column 4                  | column 5
+----------+----------+--------------------------------------------------------------------+------------------------------------+---------------------------+-----------
+ "a"      | 0        | 0x0000000000000000000000000000000000000000000000000000000000000000 | 0x00000000000000000000000000000000 | 1970-01-01T00:00:00+00:00 | +0.000000"#,
+        );
+
+        // Check struct
+        let kind: ProductType = [
+            ("bool", AlgebraicType::Bool),
+            ("str", AlgebraicType::String),
+            ("bytes", AlgebraicType::bytes()),
+            ("identity", Identity::get_type()),
+            ("connection_id", ConnectionId::get_type()),
+            ("timestamp", Timestamp::get_type()),
+            ("duration", TimeDuration::get_type()),
+        ]
+        .into();
+
+        let value = product![
+            true,
+            "This is spacetimedb".to_string(),
+            AlgebraicValue::Bytes([1, 2, 3, 4, 5, 6, 7].into()),
+            Identity::ZERO,
+            ConnectionId::ZERO,
+            Timestamp::UNIX_EPOCH,
+            TimeDuration::ZERO
+        ];
+
+        expect_psql_table(
+            &kind,
+            vec![value.clone()],
+            r#"
+ bool | str                   | bytes            | identity                                                           | connection_id                      | timestamp                 | duration
+------+-----------------------+------------------+--------------------------------------------------------------------+------------------------------------+---------------------------+-----------
+ true | "This is spacetimedb" | 0x01020304050607 | 0x0000000000000000000000000000000000000000000000000000000000000000 | 0x00000000000000000000000000000000 | 1970-01-01T00:00:00+00:00 | +0.000000"#,
+        );
+
+        // Check nested struct, tuple...
+        let kind: ProductType = [(None, AlgebraicType::product(kind))].into();
+
+        let value = product![value.clone()];
+
+        expect_psql_table(
+            &kind,
+            vec![value.clone()],
+            r#"
+ column 0
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+ (bool = true, str = "This is spacetimedb", bytes = 0x01020304050607, identity = 0x0000000000000000000000000000000000000000000000000000000000000000, connection_id = 0x00000000000000000000000000000000, timestamp = 1970-01-01T00:00:00+00:00, duration = +0.000000)"#,
+        );
+
+        let kind: ProductType = [("tuple", AlgebraicType::product(kind))].into();
+
+        let value = product![value];
+
+        expect_psql_table(
+            &kind,
+            vec![value],
+            r#"
+ tuple
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+ (0 = (bool = true, str = "This is spacetimedb", bytes = 0x01020304050607, identity = 0x0000000000000000000000000000000000000000000000000000000000000000, connection_id = 0x00000000000000000000000000000000, timestamp = 1970-01-01T00:00:00+00:00, duration = +0.000000))"#,
+        );
 
         Ok(())
     }
